# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.auditLogResponse:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: string
      reason:
        description: pointer avoids issues with NullString
        type: string
      resource_path:
        type: string
      resource_version:
        type: integer
      success:
        type: boolean
      user_email:
        type: string
    type: object
  api.createSecretRequest:
    properties:
      path:
        type: string
      ttl_seconds:
        type: integer
      value:
        type: string
    required:
    - path
    - value
    type: object
...
  api.getAuditLogsResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/api.auditLogResponse'
        type: array
    type: object
  api.getSecretResponse:
    properties:
      decrypted_value:
        type: string
      path:
        type: string
      version:
        type: integer
    type: object
...
  api.secretResponse:
    properties:
      encrypted_value:
        items:
          type: integer
        type: array
      nonce:
        items:
          type: integer
        type: array
      path:
        type: string
    type: object
...
  api.swaggerErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.updateSecretRequest:
    properties:
      value:
        type: string
    required:
    - value
    type: object
  api.updateSecretResponse:
    properties:
      encrypted_value:
        items:
          type: integer
        type: array
      nonce:
        items:
          type: integer
        type: array
      path:
        type: string
      version:
        type: integer
    type: object
...
host: localhost:9090
info:
  contact:
    email: mishrayashaswikumar@gmail.com
    name: Yashaswi Kumar Mishra
  description: Secure secret management system in Go
  title: Vaultify API
  version: "1.0"
paths:
  /audit/logs:
    get:
      consumes:
      - application/json
      description: Fetch audit logs for the authenticated user with optional filters
      parameters:
      - description: Action filter (e.g., CREATE_SECRET)
        in: query
        name: action
        type: string
      - description: Resource path filter (e.g., /vault/secrets/foo)
        in: query
        name: path
        type: string
      - description: Resource version filter
        in: query
        name: version
        type: integer
      - description: Success status filter (true/false)
        in: query
        name: success
        type: boolean
      - description: Start date in RFC3339 or YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: End date in RFC3339 or YYYY-MM-DD
        in: query
        name: to
        type: string
      - description: Limit number of results (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of audit logs
          schema:
            items:
              $ref: '#/definitions/api.getAuditLogsResponse'
            type: array
        "400":
          description: Invalid query parameter
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "404":
          description: No logs found
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit logs
      tags:
      - Audit
...
  /secrets:
    post:
      consumes:
      - application/json
      description: Encrypts and stores a secret with optional TTL, linked to the authenticated
        user. The encrypted secret is signed with an HMAC signature to ensure integrity
        and prevent tampering.
      parameters:
      - description: Secret creation request
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/api.createSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.secretResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new secret
      tags:
      - Secrets
  /secrets/{path}:
    get:
      description: Fetches and decrypts the secret. If version is not specified, retrieves
        the latest. Verifies HMAC to ensure integrity.
      parameters:
      - description: Secret path
        in: path
        name: path
        required: true
        type: string
      - description: Secret version (optional)
        in: query
        name: version
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.getSecretResponse'
        "401":
          description: Unauthorized or HMAC verification failed
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "404":
          description: Secret not found
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve a secret by path and optional version
      tags:
      - Secrets
    put:
      consumes:
      - application/json
      description: Encrypts new secret value, verifies existing HMAC to prevent tampering,
        then creates a new secret version signed with a fresh HMAC.
      parameters:
      - description: Secret path
        in: path
        name: path
        required: true
        type: string
      - description: New secret value
        in: body
        name: updateSecretRequest
        required: true
        schema:
          $ref: '#/definitions/api.updateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.updateSecretResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "401":
          description: Unauthorized or HMAC verification failed
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "404":
          description: Secret not found
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing secret by creating a new version
      tags:
      - Secrets
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-paseto-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"